requirements.txt
argparse
pyserial
sounddevice
numpy
pynput
